[KAN-233] refactor(billing): 청구 금액 및 개수 미리 계산해서 쿼리 성능 개선

+++ server/src/main/java/kr/or/kosa/cmsplusmain/domain/billing/dto/response/BillingListItemRes.java
@@ -20,12 +20,12 @@ public class BillingListItemRes {
 	private final LocalDate billingDate;
 	@Setter
 	private String firstProductName;
-	private final Long totalProductCount;
+	private final Integer totalProductCount;
 
 	@QueryProjection
 	public BillingListItemRes(Long billingId, String memberName, String memberPhone, Long billingPrice,
 		BillingStatus billingStatus, PaymentType paymentType, LocalDate billingDate, String firstProductName,
-		Long totalProductCount) {
+		Integer totalProductCount) {
 		this.billingId = billingId;
 		this.memberName = memberName;
 		this.memberPhone = memberPhone;
